Dev --> Master pr March 26, 2025 v8.1.0

+++ deploy-nrrd.sh
@@ -18,47 +18,61 @@ show_spinner() {
 # Redirect all output to a log file (silent in terminal)
 exec > deploy-nrrd.log 2>&1
 
-echo -n "Enter the cloud.gov space you are targeting: " > /dev/tty
-read space < /dev/tty
+# handle preview deployments
+echo -n "Is this deployment for preview? (y/n) " > /dev/tty
+read is_preview < /dev/tty
 
-# Check if cf is logged in
-echo "Checking to see if you're logged in to cf..." > /dev/tty
-cf apps > /dev/null 2>&1 &
-pid=$!
-show_spinner $pid
-wait $pid
-if [ $? -ne 0 ]; then
-  echo -n "Enter a temporary authentication code ( Get one at https://login.fr.cloud.gov/passcode ): " > /dev/tty
-  read -s sso_passcode < /dev/tty
-  echo > /dev/tty
-  echo "Logging in to cf..." > /dev/tty
-  cf login -a api.fr.cloud.gov -o doi-onrr -s $space --sso-passcode $sso_passcode &
+if [ "$is_preview" = "y" ]; then
+  echo -n "Enter your Github branch to deploy to: " > /dev/tty
+  read circle_branch < /dev/tty
+
+  hasura_uri="https://hasura-sandbox.app.cloud.gov/v1/graphql"
+  stage="nrrd-preview"
+else
+  echo -n "Enter the cloud.gov space you are targeting: " > /dev/tty
+  read space < /dev/tty
+
+  # Check if cf is logged in
+  echo "Checking to see if you're logged in to cf..." > /dev/tty
+  cf apps > /dev/null 2>&1 &
   pid=$!
   show_spinner $pid
-  wait $pid || { echo "cf login failed. Check the log file. Exiting..." > /dev/tty; exit 1; }
-else
-  echo "Targeting cf space..." > /dev/tty
-  cf target -s $space &
-  pid=$!
-  show_spinner $pid
-  wait $pid || { echo "cf target failed. Check the log file. Exiting..." > /dev/tty; exit 1; }
-fi
+  wait $pid
+  if [ $? -ne 0 ]; then
+    echo -n "Enter a temporary authentication code ( Get one at https://login.fr.cloud.gov/passcode ): " > /dev/tty
+    read -s sso_passcode < /dev/tty
+    echo > /dev/tty
+    echo "Logging in to cf..." > /dev/tty
+    cf login -a api.fr.cloud.gov -o doi-onrr -s $space --sso-passcode $sso_passcode &
+    pid=$!
+    show_spinner $pid
+    wait $pid || { echo "cf login failed. Check the log file. Exiting..." > /dev/tty; exit 1; }
+  else
+    echo "Targeting cf space..." > /dev/tty
+    cf target -s $space &
+    pid=$!
+    show_spinner $pid
+    wait $pid || { echo "cf target failed. Check the log file. Exiting..." > /dev/tty; exit 1; }
+  fi
 
-# Validate space and set branch
-if [ "$space" = "prod" ]; then
-  circle_branch="master"
-  app="nrrd"
-elif [ "$space" = "dev" ]; then
-  circle_branch="dev"
-  app="dev-nrrd"
-else
-  echo "Unknown space: $space"
-  exit 1
+  # Validate space and set branch
+  if [ "$space" = "prod" ]; then
+    circle_branch="master"
+    app="nrrd"
+  elif [ "$space" = "dev" ]; then
+    circle_branch="dev"
+    app="dev-nrrd"
+  else
+    echo "Unknown space: $space"
+    exit 1
+  fi
+  hasura_uri="https://hasura-${space}.app.cloud.gov/v1/graphql"
+  stage="${space}"
 fi
 
 export CIRCLE_BRANCH="${circle_branch}"
-export CIRCLE_STAGE="${space}"
-export GATSBY_HASURA_URI="https://hasura-${space}.app.cloud.gov/v1/graphql"
+export CIRCLE_STAGE="${stage}"
+export GATSBY_HASURA_URI="${hasura_uri}"
 
 # Install npm deps if they aren't installed
 if [ ! -d "node_modules" ]; then
@@ -95,7 +109,13 @@ mv ./public/~partytown ./public/_partytown
 
 # Deploying
 echo "Deploying..." > /dev/tty
-cf push $app -f manifest.yml &
+
+if ["$is_preview" = "n"]; then
+  cf push $app -f manifest.yml &
+else 
+  aws s3 sync public "s3://${AWS_BUCKET_NAME}/sites/${CIRCLE_BRANCH}/" &
+fi
+
 pid=$!
 show_spinner $pid
 wait $pid || { echo "Deployment failed. Check the log file. Exiting..." > /dev/tty; exit 1; }
